If loops added for actions

+++ controllers/CompanyController.php
@@ -62,13 +62,18 @@ class CompanyController extends Controller
      */
     public function actionIndex()
     {
-        $searchModel = new SearchCompany();
-        $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
+        if (\Yii::$app->user->can('indexCompany')){
+            $searchModel = new SearchCompany();
+            $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
-        return $this->render('index', [
-            'searchModel' => $searchModel,
-            'dataProvider' => $dataProvider,
-        ]);
+            return $this->render('index', [
+                'searchModel' => $searchModel,
+                'dataProvider' => $dataProvider,
+            ]);
+        }else{
+            throw new \yii\web\ForbiddenHttpException;
+        }
+        
     }
 
     /**
@@ -78,12 +83,16 @@ class CompanyController extends Controller
      */
     public function actionView($id)
     {
-        $model = $this->findModel($id);
-        $orders = Orders::find()->where(['company_id' => $model->company_id])->groupBy('order_number')->all();
-        return $this->render('view', [
-            'model' => $model,
-            'orders' => $orders
-        ]);
+        if (\Yii::$app->user->can('ViewCompany')){
+            $model = $this->findModel($id);
+            $orders = Orders::find()->where(['company_id' => $model->company_id])->groupBy('order_number')->all();
+            return $this->render('view', [
+                'model' => $model,
+                'orders' => $orders
+            ]);
+        }else{
+            throw new \yii\web\ForbiddenHttpException;
+        }
     }
 
     /**
@@ -93,20 +102,24 @@ class CompanyController extends Controller
      */
     public function actionCreate()
     {
-        $model = new Company();
-        $user = new Users();
+        if (\Yii::$app->user->can('createCompany')){
+            $model = new Company();
+            $user = new Users();
 
-        if ($model->load(Yii::$app->request->post()) && $user->load(Yii::$app->request->post())) {
-            $user->password = Yii::$app->getSecurity()->generatePasswordHash($user->password);
-            $user->save();
-            $model->user_id = $user->user_id;
-            $model->save();
-            return $this->redirect(['view', 'id' => $model->company_id]);
-        } else {
-            return $this->render('create', [
-                'model' => $model,
-                'user' => $user
-            ]);
+            if ($model->load(Yii::$app->request->post()) && $user->load(Yii::$app->request->post())) {
+                $user->password = Yii::$app->getSecurity()->generatePasswordHash($user->password);
+                $user->save();
+                $model->user_id = $user->user_id;
+                $model->save();
+                return $this->redirect(['view', 'id' => $model->company_id]);
+            } else {
+                return $this->render('create', [
+                    'model' => $model,
+                    'user' => $user
+                ]);
+            }
+        }else{
+            throw new \yii\web\ForbiddenHttpException;
         }
     }
 
@@ -118,14 +131,18 @@ class CompanyController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
+        if (\Yii::$app->user->can('updateCompany')){
+            $model = $this->findModel($id);
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->company_id]);
-        } else {
-            return $this->render('update', [
-                'model' => $model,
-            ]);
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->company_id]);
+            } else {
+                return $this->render('update', [
+                    'model' => $model,
+                ]);
+            }
+        }else{
+            throw new \yii\web\ForbiddenHttpException;
         }
     }
 
@@ -137,9 +154,12 @@ class CompanyController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
-
-        return $this->redirect(['index']);
+        if (\Yii::$app->user->can('deleteCompany')){
+            $this->findModel($id)->delete();
+            return $this->redirect(['index']);
+        }else{
+            throw new \yii\web\ForbiddenHttpException;
+        }
     }
 
     /**
