+++ controllers/OrdersController.php
@@ -40,13 +40,17 @@ class OrdersController extends Controller
      */
     public function actionIndex()
     {
-        $searchModel = new SearchOrders();
-        $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
-        
-        return $this->render('index', [
-            'searchModel' => $searchModel,
-            'dataProvider' => $dataProvider,
-        ]);
+        if (\Yii::$app->user->can('indexOrders')){
+            $searchModel = new SearchOrders();
+            $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
+            
+            return $this->render('index', [
+                'searchModel' => $searchModel,
+                'dataProvider' => $dataProvider,
+            ]);
+        }else{
+            throw new \yii\web\ForbiddenHttpException;
+        }
     }
 
     /**
@@ -56,15 +60,19 @@ class OrdersController extends Controller
      */
     public function actionView($id)
     {
-        $model = $this->findModel($id);
-        $plots = OrderDetails::find()->where(['order_id' => $id]);
-        $dataProvider = new ActiveDataProvider([
-            'query' => $plots,
-        ]);
-        return $this->render('view', [
-            'model' =>  $model,
-            'plots' => $dataProvider,
-        ]);
+        if (\Yii::$app->user->can('viewOrders')){
+            $model = $this->findModel($id);
+            $plots = OrderDetails::find()->where(['order_id' => $id]);
+            $dataProvider = new ActiveDataProvider([
+                'query' => $plots,
+            ]);
+            return $this->render('view', [
+                'model' =>  $model,
+                'plots' => $dataProvider,
+            ]);
+        }else{
+            throw new \yii\web\ForbiddenHttpException;
+        }
     }
 
     /**
@@ -74,29 +82,33 @@ class OrdersController extends Controller
      */
     public function actionCreate()
     {
-        $model = new Orders();
-        $company = Company::find()->all();
-        $area = Area::find()->all();
-        $orderDetails = new OrderDetails();
-        if ($model->load(Yii::$app->request->post()) && $orderDetails->load(Yii::$app->request->post())) {
-            $model->save();
-            for($i = 0; $i < sizeof($orderDetails->plot_id); $i++){
-                $detail = new OrderDetails();
-                $plot = new Plot();
-                $plot->name = $orderDetails->plot_id[$i];
-                $plot->save(false);
-                $detail->order_id = $model->order_id;
-                $detail->plot_id = $plot->plot_id;
-                $detail->save(false);
+        if (\Yii::$app->user->can('createOrders')){
+            $model = new Orders();
+            $company = Company::find()->all();
+            $area = Area::find()->all();
+            $orderDetails = new OrderDetails();
+            if ($model->load(Yii::$app->request->post()) && $orderDetails->load(Yii::$app->request->post())) {
+                $model->save();
+                for($i = 0; $i < sizeof($orderDetails->plot_id); $i++){
+                    $detail = new OrderDetails();
+                    $plot = new Plot();
+                    $plot->name = $orderDetails->plot_id[$i];
+                    $plot->save(false);
+                    $detail->order_id = $model->order_id;
+                    $detail->plot_id = $plot->plot_id;
+                    $detail->save(false);
+                }
+                return $this->redirect(['orders/index']);
+            } else {
+                return $this->render('create', [
+                    'model' => $model,
+                    'company' => $company,
+                    'area' => $area,
+                    'orderDetails' => $orderDetails,
+                ]);
             }
-            return $this->redirect(['orders/index']);
-        } else {
-            return $this->render('create', [
-                'model' => $model,
-                'company' => $company,
-                'area' => $area,
-                'orderDetails' => $orderDetails,
-            ]);
+        }else{
+            throw new \yii\web\ForbiddenHttpException;
         }
     }
 
@@ -108,29 +120,33 @@ class OrdersController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model =  Orders::find()->where(['order_id' => $id])->all();
-        $company = Company::find()->all();
-        $area = Area::find()->all();
-        $orderDetails = OrderDetails::find()->where(['order_id' => $id])->all();
-        if ($model->load(Yii::$app->request->post()) && $orderDetails->load(Yii::$app->request->post())) {
-            $model->save();
-            for($i = 0; $i < sizeof($orderDetails->plot_id); $i++){
-                $detail = new OrderDetails();
-                $plot = new Plot();
-                $plot->name = $orderDetails->plot_id[$i];
-                $plot->save(false);
-                $detail->order_id = $model->order_id;
-                $detail->plot_id = $plot->plot_id;
-                $detail->save(false);
+        if (\Yii::$app->user->can('updateOrders')){
+            $model =  Orders::find()->where(['order_id' => $id])->all();
+            $company = Company::find()->all();
+            $area = Area::find()->all();
+            $orderDetails = OrderDetails::find()->where(['order_id' => $id])->all();
+            if ($model->load(Yii::$app->request->post()) && $orderDetails->load(Yii::$app->request->post())) {
+                $model->save();
+                for($i = 0; $i < sizeof($orderDetails->plot_id); $i++){
+                    $detail = new OrderDetails();
+                    $plot = new Plot();
+                    $plot->name = $orderDetails->plot_id[$i];
+                    $plot->save(false);
+                    $detail->order_id = $model->order_id;
+                    $detail->plot_id = $plot->plot_id;
+                    $detail->save(false);
+                }
+                return $this->redirect(['orders/index']);
+            } else {
+                return $this->render('update', [
+                    'model' => $model,
+                    'company' => $company,
+                    'area' => $area,
+                    'orderDetails' => $orderDetails,
+                ]);
             }
-            return $this->redirect(['orders/index']);
-        } else {
-            return $this->render('update', [
-                'model' => $model,
-                'company' => $company,
-                'area' => $area,
-                'orderDetails' => $orderDetails,
-            ]);
+        }else{
+            throw new \yii\web\ForbiddenHttpException;
         }
     }
 
@@ -142,9 +158,12 @@ class OrdersController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
-
-        return $this->redirect(['index']);
+        if (\Yii::$app->user->can('deleteOrders')){
+            $this->findModel($id)->delete();
+            return $this->redirect(['index']);
+        }else{
+            throw new \yii\web\ForbiddenHttpException;
+        }
     }
 
     /**
